relocates ape fixtures to local conftest
